encapsulate type reference comparison
